Eliminate more warnings on Borland C.
SVN-Revision: 1774

+++ tar/test/main.c
@@ -1061,7 +1061,7 @@ assertion_file_nlinks(const char *file, int line,
 
 	assertion_count(file, line);
 	r = my_GetFileInformationByName(pathname, &bhfi);
-	if (r != 0 && bhfi.nNumberOfLinks == nlinks)
+	if (r != 0 && bhfi.nNumberOfLinks == (DWORD)nlinks)
 		return (1);
 	failure_start(file, line, "File %s has %d links, expected %d",
 	    pathname, bhfi.nNumberOfLinks, nlinks);
